Unit tests: implement `ModelTestCase` test case class
This reduces duplication of the example code.
