+++ tests/reconstruction_gaussian_model.py
@@ -1,14 +1,9 @@
-# Core imports
-import unittest
-
-# Library imports
-import numpy as np
-
 # Package imports
 from ..reconstruction.Gaussian_Model import Gaussian_Model
 from ..settings.Arguments import Arguments
+from reconstruction_model import ModelTestCase
 
-class TestReconstructionGaussianModel(unittest.TestCase):
+class TestReconstructionGaussianModel(ModelTestCase):
     def setUp(self):
         super(TestReconstructionGaussianModel, self).setUp()
 
@@ -26,23 +21,8 @@ class TestReconstructionGaussianModel(unittest.TestCase):
         self.assertEqual(self.model.type, "reconstruction_gaussian_model")
 
     def test_assign(self):
-        def distance(x, y):
-            # Helper function to calculate the distance from location
-            # (0, 0) to the pixel center location of pixel (x, y).
-            return np.sqrt((x + 0.5) ** 2 + (y + 0.5) ** 2)
-
-        # The grid contains 16 pixels (four by four). The link goes from
-        # location (0, 0) to location (4, 4). Location (0, 0) is located
-        # in the top left corner and location (4, 4) is located in the
-        # bottom right corner of the grid.
-        length = np.sqrt(4 ** 2 + 4 ** 2)
-        source_distances = np.array([
-            distance(0, 0), distance(1, 0), distance(2, 0), distance(3, 0),
-            distance(0, 1), distance(1, 1), distance(2, 1), distance(3, 1),
-            distance(0, 2), distance(1, 2), distance(2, 2), distance(3, 2),
-            distance(0, 3), distance(1, 3), distance(2, 3), distance(3, 3)
-        ]).reshape(4, 4)
-        destination_distances = np.flipud(np.fliplr(source_distances))
+        length, source_distances, destination_distances = \
+            super(TestReconstructionGaussianModel, self).test_assign()
 
         # The assigned weights must be generated by the Gaussian function.
         weights = self.model.assign(length, source_distances, destination_distances)
